+++ docs/src/demos/Nodes/TaskList/index.vue
@@ -37,7 +37,7 @@ export default {
         TaskItem,
       ],
       content: `
-        <ul data-type="task_list">
+        <ul data-type="taskList">
           <li data-type="taskItem" data-checked="true">A list item</li>
           <li data-type="taskItem" data-checked="false">And another one</li>
         </ul>