+++ packages/core/src/helpers/generateHTML.ts
@@ -1,9 +1,9 @@
 import { Node } from 'prosemirror-model'
 import getSchema from './getSchema'
 import getHTMLFromFragment from './getHTMLFromFragment'
-import { Extensions } from '../types'
+import { Extensions, JSONContent } from '../types'
 
-export default function generateHTML(doc: object, extensions: Extensions): string {
+export default function generateHTML(doc: JSONContent, extensions: Extensions): string {
   const schema = getSchema(extensions)
   const contentNode = Node.fromJSON(schema, doc)