+++ packages/extension-bubble-menu/src/bubble-menu-plugin.ts
@@ -1,4 +1,9 @@
-import { Editor, posToDOMRect, isNodeSelection } from '@tiptap/core'
+import {
+  Editor,
+  posToDOMRect,
+  isTextSelection,
+  isNodeSelection,
+} from '@tiptap/core'
 import { EditorState, Plugin, PluginKey } from 'prosemirror-state'
 import { EditorView } from 'prosemirror-view'
 import tippy, { Instance, Props } from 'tippy.js'
@@ -34,6 +39,7 @@ export class BubbleMenuView {
     this.element = element
     this.view = view
     this.element.addEventListener('mousedown', this.mousedownHandler, { capture: true })
+    this.view.dom.addEventListener('dragstart', this.dragstartHandler)
     this.editor.on('focus', this.focusHandler)
     this.editor.on('blur', this.blurHandler)
     this.createTooltip(tippyOptions)
@@ -42,6 +48,10 @@ export class BubbleMenuView {
 
   mousedownHandler = () => {
     this.preventHide = true
+  }
+
+  dragstartHandler = () => {
+    this.hide()
   }
 
   focusHandler = () => {
@@ -88,7 +98,7 @@ export class BubbleMenuView {
       return
     }
 
-    const { empty, $anchor, ranges } = selection
+    const { empty, ranges } = selection
 
     // support for CellSelections
     const from = Math.min(...ranges.map(range => range.$from.pos))
@@ -97,7 +107,10 @@ export class BubbleMenuView {
     // Sometime check for `empty` is not enough.
     // Doubleclick an empty paragraph returns a node size of 2.
     // So we check also for an empty text size.
-    if (empty || !$anchor.parent.textContent) {
+    const isEmptyTextBlock = !doc.textBetween(from, to).length
+      && isTextSelection(view.state.selection)
+
+    if (empty || isEmptyTextBlock) {
       this.hide()
 
       return
@@ -131,6 +144,7 @@ export class BubbleMenuView {
   destroy() {
     this.tippy.destroy()
     this.element.removeEventListener('mousedown', this.mousedownHandler)
+    this.view.dom.removeEventListener('dragstart', this.dragstartHandler)
     this.editor.off('focus', this.focusHandler)
     this.editor.off('blur', this.blurHandler)
   }