+++ packages/html/src/generateHTML.ts
@@ -1,8 +1,8 @@
-import { Extensions, getSchema } from '@tiptap/core'
+import { Extensions, getSchema, JSONContent } from '@tiptap/core'
 import { Node } from 'prosemirror-model'
 import getHTMLFromFragment from './getHTMLFromFragment'
 
-export default function generateHTML(doc: object, extensions: Extensions): string {
+export default function generateHTML(doc: JSONContent, extensions: Extensions): string {
   const schema = getSchema(extensions)
   const contentNode = Node.fromJSON(schema, doc)