+++ packages/extension-dropcursor/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.16](https://github.com/ueberdosis/tiptap/compare/@tiptap/extension-dropcursor@2.0.0-beta.15...@tiptap/extension-dropcursor@2.0.0-beta.16) (2021-06-15)
+
+**Note:** Version bump only for package @tiptap/extension-dropcursor
+
+
+
+
+
+# [2.0.0-beta.15](https://github.com/ueberdosis/tiptap/compare/@tiptap/extension-dropcursor@2.0.0-beta.14...@tiptap/extension-dropcursor@2.0.0-beta.15) (2021-06-14)
+
+**Note:** Version bump only for package @tiptap/extension-dropcursor
+
+
+
+
+
 # [2.0.0-beta.14](https://github.com/ueberdosis/tiptap/compare/@tiptap/extension-dropcursor@2.0.0-beta.13...@tiptap/extension-dropcursor@2.0.0-beta.14) (2021-05-27)
 
 **Note:** Version bump only for package @tiptap/extension-dropcursor